requirements.txt
Pillow==8.0.1
tensorflow-gpu==1.15.0